requirements.txt
# API Integration
tweepy>=4.14.0
requests>=2.31.0
aiohttp>=3.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0

# Logging & Monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Security
cryptography>=41.0.0
keyring>=24.0.0

# Utilities
click>=8.1.0
schedule>=1.2.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
responses>=0.23.0
